[IMP] pos_price_to_weight: conversion to ESM
